Refactor environment setup and passport commands
Modify environment file handling and update passport commands.

+++ auto_builder_sh.sh
@@ -1,14 +1,20 @@
 rm -rf package
 rm storage/install.zip
 rm storage/update.zip
+
 rsync -av --exclude '.env' --exclude 'auto_builder.sh' --exclude 'vendor/' --exclude '.git/' --exclude '.idea/' ./ ./package
 
 cd package
+
 composer update --no-interaction --prefer-dist
-cp .env.example .env
-php artisan key:generate
+
+if [ ! -f .env ]; then
+  cp .env.example .env
+  php artisan key:generate
+fi
+
 php artisan passport:keys --skip-if-exists
-php artisan passport:install --copy --force
+php artisan passport:client --personal --no-interaction
 
 mkdir -p storage/framework/{sessions,views,cache}
 chmod -R 775 storage/framework
@@ -19,7 +25,6 @@ zip -r ../storage/install.zip .
 
 php artisan prepare:updatable
 rm -rf installation
-rm -rf storage
 rm .env
 zip -r ../storage/update.zip .
